refactor menu styles
