--- src/base-components/index.ts
@@ -1,1 +1,0 @@
-export { MenuBackdrop, MenuContainer, MenuList } from './styled/Menu';
