+++ src/sidebar/Menu/ContextMenu/useContextMenu.ts
@@ -1,13 +1,13 @@
 import { useCallback, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 
-import { menuWidthInPx } from '../../base-components/styled/Menu';
-import sessionSlice from '../../store/slices/session';
-import useWindowDimensions from '../../utils/hooks/useWindowDimensions';
+import sessionSlice from '../../../store/slices/session';
+import useWindowDimensions from '../../../utils/hooks/useWindowDimensions';
 
 // TODO find a more robust way to determine menu height
 // find DOM node in useLayoutEffect and use getBoundingClientRect?
 const contextMenuHeight = 64; // 2 menu items, each 32px
+const menuWidthInPx = 200; // TODO use custom property?
 
 export const useContextMenu = (containerRef: React.RefObject<HTMLDivElement>) => {
     const { height, width } = useWindowDimensions();
