+++ src/sidebar/Menu/MenuListItem/menu-list-item.css
@@ -1,0 +1,26 @@
+.menu__list-item {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    list-style: none;
+
+    padding: 6.5px;
+    padding-left: 24px;
+
+    &:hover {
+        background-color: var(--menu-hover-background-color);
+        color: var(--menu-hover-text-color);
+    }
+
+    &:has(.menu__list-item-icon) {
+        padding-left: 4px;
+    }
+}
+
+.menu__list-item-icon {
+    /* center the icon */
+    width: 24px;
+    height: 18px;
+    padding-left: 4px;
+    margin-right: 2px;
+}
