+++ src/sidebar/Menu/MoreMenu/more-menu.css
@@ -1,0 +1,7 @@
+.more-menu__import-input {
+    /* do not take up any space */
+    display: block;
+    height: 0;
+
+    visibility: hidden;
+}
