--- src/sidebar/Menu/index.ts
@@ -1,3 +1,0 @@
-import Menu from './Menu';
-
-export default Menu;
