+++ src/sidebar/Menu/menu.css
@@ -1,3 +1,38 @@
+@import './MenuListItem/menu-list-item.css';
+@import './MoreMenu/more-menu.css';
+
+.menu__backdrop {
+    position: absolute;
+    top: 0;
+    width: 100%;
+    height: 100%;
+}
+.menu__backdrop--hidden {
+    z-index: -1;
+    visibility: hidden;
+}
+
+.menu__container {
+    position: absolute;
+    top: var(--menu-anchor-top);
+    left: var(--menu-anchor-left);
+
+    width: var(--menu-width);
+    padding: 1px;
+
+    border: 1px solid var(--menu-border-color);
+    border-radius: 2px;
+    background-color: var(--sidebar-background-color);
+    color: var(--sidebar-tex-color);
+}
+
+.menu__list {
+    padding: 0;
+    margin: 0;
+
+    list-style: none;
+}
+
 .menu__divider {
     border: none;
     border-bottom: 1px solid var(--menu-border-color);
